Upgraded LLM Orchestrator with dynamic state, reasoning prompts and workflow logic overall
