+++ src/floki/workflow/orchestrators/__init__.py
@@ -1,3 +1,4 @@
+from .base import OrchestratorServiceBase
 from .llm import LLMOrchestrator
 from .random import RandomOrchestrator
 from .roundrobin import RoundRobinOrchestrator
